added "if" statement to only load NPPSD once, not thrice

+++ Benthic_Nearshore_Drivers/Data_Cleaning_Scripts_DMX_Benthic_Nearshore/HADUAbun_DMX_Benthic.R
@@ -22,28 +22,30 @@ library(RODBC)
 #############
 # Harlequin Duck Abundance
 
-NPPSDzipd <- tempfile()
-download.file("http://alaska.usgs.gov/data/nppsd/nppsd.zip", NPPSDzipd, mode="wb")
-NPPSDzip_L <- unzip(NPPSDzipd, list=TRUE)
+if(!exists("DATA_OBS")|!exists("LOCATION")) {
+                      NPPSDzipd <- tempfile()
+                      download.file("http://alaska.usgs.gov/data/nppsd/nppsd.zip", NPPSDzipd, mode="wb")
+                      NPPSDzip_L <- unzip(NPPSDzipd, list=TRUE)
 
-NPPSDzip_DB <- NPPSDzip_L[grep(".mdb", NPPSDzip_L$Name),]$Name   # creates list of files I want
-UNz <- unzip(NPPSDzipd, NPPSDzip_DB) # unzip the two mdb files (makes list I think)
+                      NPPSDzip_DB <- NPPSDzip_L[grep(".mdb", NPPSDzip_L$Name),]$Name   # creates list of files I want
+                      UNz <- unzip(NPPSDzipd, NPPSDzip_DB) # unzip the two mdb files (makes list I think)
 
-mdb_table_list <- function(file_list){
-                  # for every mdb file in the list, do the following 
-                  conn <- odbcConnectAccess2007(path.expand(file_list)) # establish a connection
-                  table_list <- subset(sqlTables(conn), TABLE_TYPE=="TABLE") # lists tables in DB
-                  return(table_list)
-                  }
+                      mdb_table_list <- function(file_list){
+                                       # for every mdb file in the list, do the following 
+                                        conn <- odbcConnectAccess2007(path.expand(file_list)) # establish a connection
+                                        table_list <- subset(sqlTables(conn), TABLE_TYPE=="TABLE") # lists tables in DB
+                                        return(table_list)
+                                        }
 
-lapply(UNz, mdb_table_list)  # running the function over the two .mdb files
+                      lapply(UNz, mdb_table_list)  # running the function over the two .mdb files
 
-conn <- odbcConnectAccess2007(path.expand("./NPPSD_Back_v2.mdb")) # establish a connection
-DATA_OBS <- sqlFetch(conn,"tbl_DATA_OBS")  # read in a table
-LOCATION <- sqlFetch(conn,"tbl_LOCATION") 
+                      conn <- odbcConnectAccess2007(path.expand("./NPPSD_Back_v2.mdb")) # establish a connection
+                      DATA_OBS <- sqlFetch(conn,"tbl_DATA_OBS")  # read in a table
+                      LOCATION <- sqlFetch(conn,"tbl_LOCATION") 
 
-close(conn) 
-unlink(NPPSDzipd)
+                      close(conn) 
+                      unlink(NPPSDzipd)
+                      }
 
 # Filter and Sort data       # Density is Number of Birds per km2
 EPWS <- c("60.32663-146.6541" , "60.32763-146.6633" , "60.32983-146.664", "60.34233-146.67664",
